Gave two recipe webpages proper headers and structure.
Both webpages previously had only <body> code and no header and doctype or html tags so I added those as well as a link to the .css style sheet for the whole site.

This was done because I thought it was what was keeping the <link> to the stylesheet from working but it turns out I just had to wrong relative address. I kept it because its not hurting anything.